backend/requirements.txt
asgiref==3.6.0
attrs==23.1.0
backports.zoneinfo==0.2.1
Django==4.2
django-allauth==0.54.0
django-cors-headers==3.14.0
django-filter==23.1
django-rest-auth==0.9.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
drf-standardized-errors==0.12.5
importlib-resources==5.12.0
inflection==0.5.1
jsonschema==4.17.3
Pillow==9.5.0
pkgutil-resolve-name==1.3.10
psycopg2-binary==2.9.6
PyJWT==2.6.0
pyotp==2.8.0
pyrsistent==0.19.3
python-decouple==3.8
python-dotenv==1.0.0
python-http-client==3.3.7
pytz==2023.3
PyYAML==6.0
sendgrid==6.10.0
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
uritemplate==4.1.1
zipp==3.15.0

pillow
